msm8996-common: rootdir: Set DAC sysfs node ownership if available
V20 has support for the ESS ES9218 DAC and its new revanped panel
now requires the usage of sysfs nodes for many of the DAC's
settings. So add a check into init.target to see if the target
device has support for that DAC and set all sysfs nodes' ownership
to system accordingly.

Change-Id: Ie2de67751f6a6ea2133190ae74449966712262c6